replace registrySettings with fileSettings
update integration tests to work with file based settings
instead of providing settings via the registry which got
removed we just inject the filesettings where the registry
code would have put them eventually. this should work as
a stand in replacement for the registry and have the equivalent
outcome from the agents perspective.

remove config drive setup

switching to filesettings makes this obsolete.