- add ipv6 test for nats fw
we didn't test the nats iptables rules for nats fw before, this simply
adds them to add coverage.

use defaults in the test-environment if no value is set. To be able to
test for ipv6 the mbus url cannot be hardcoded to an ipv4 address in the
test env. Only set the defaults in the test-env if the test-setup didn't
provide values.

- update sequence of test setups for ephemeral disks

move setup to the outer context since it's required across tests and not
just in that specific test.