Handles "system" & "all" types in fetch log action
This commit updates the Bosh Agent to support the "system" and "all" log
types that are now (or will soon be) supported by the Bosh CLI. This
work is backwards-compatible with versions of the Bosh CLI that do not
support these new log types.

The "system" log type fetches the logs from "/var/log". System logs are
placed into a "/var/log" directory inside the tarball. This is done to
generally avoid confusion as to the source of these logs, but especially
when the "all" log type is requested. Do note that if the "all" log type
is requested, and the VM has a job named "var" deployed, then the system
logs will be placed into a "log" subdirectory in the "var" job's log
directory. For now, this seems to be a minor and unlikely problem, so
avoiding this doesn't seem to be worth the hassle.

The "all" log type is a pseudo-type constructed by the Bosh CLI, which
fetches all currently-known log types; "job" (the historical (and
current) default when no log type is specified), "agent", and "system".

NOTE: Requests for the "system" log type will be sliently ignored when
      the Agent is running on an OS that is not Linux. Currently, only
      Windows and Linux are supported OSs. If we ever support another
      Unix-alike OS (or figure out how to do something sensible and
      useful on Windows), then the code that handles the "system" log
      type will need to be updated.

NOTE: The fetch log action can handle combinations of log types that
      are not supported by the Bosh CLI. This should generally be
      harmless, and probably makes it easier to extend in the future.
      We also do no deduplication of the log types passed to the action.
      A malicious user could take advantage of this to cause the Agent
      to consume large amounts of RAM or disk space (whether in the
      blobstore, or on the Agent's disk), but if a malicious user has
      the credentials required to talk to the Agent, they can already
      get up to significant mischief.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com
Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>