Retry umount when cleaning up mounts
We believe there's a race condition between `fuser -km MOUNT` killing a
process and being able to unmount the mount point using umount. When we
added a retry to umount on a debug branch, it was able to successfully
unmount after a few retries. The 100 chosen here is arbitrary; even if
100 attempts are made, it should be fast.

We've also fixed debug logging when the RunCommand method fails;
previously, it was printing the string literal instead of interpolating
the command and error, i.e.
```
[COMMAND FAILED TO EXECUTE: %s ERROR: %s
] 2023/04/03 22:34:19 DEBUG - sudo fuser -km /opt%!(EXTRA *ssh.ExitError=Process exited with status 1)
```

Signed-off-by: Brian Upton <bupton@vmware.com>