Add handling for disk IDs with prefix
Some IaaS provider have disk ID with prefix which is separated
by a hyphen character. However the symlink doesn't have the prefix
and the hyphen character that is why the resolution doesn't work in this
case. This pr removes the prefix with the hyphen from the disk ID before
the resolution to provide a solution for this case