The agent is not currently capable of handling multiple async tasks in parallel.
When the second task is received, StartTask tries to put the task
onto the taskChan of the async task service but that channel is
already full with the current processing task. This prevents StartTask
from returning, which prevents the dispatcher from returning, which
prevents all incoming actions from being processed by the handler.

Once the handler and dispatcher are blocked, the agent is unable to
process incoming get_task requests, which ultimately leads to the
director timing out.

This commit changes the taskChan to a length of 5. This won't actually
let the agent process multiple async tasks at the same time, but it will
accept them and store them on the channel without blocking the dispatcher
or the handler. This allow the get_tasks for the current async tasks
to continue to be processed. Once the task it complete, the additional
queued tasks can be processed.

You can reproduce the problem state by having a VM perform a long
running task such as a pre-start or drain script. Then perform a "logs"
action on the VM which calls the "fetch_logs" async action.

Even with this change, it will still be possible to get the VM into a
broken state by performing more than 5 concurrent async actions at the same
time. Such as multiple fetch logs attempts; since the first one won't work
if the agent is processing a different task, they may cancel and retry it
which will just queue up additional tasks. But this changes makes it
less likely to cause a deploy to fail by attempting to perform a second
task at the same time.