Bump WinSW to forked version based on master branch
We found that versions of WinSW > v2.11.0 had the behavior that if a
wrapped service exits 0, WinSW treats that as a success and stops the
service.  In the context of BOSH, we want any exit to result in the
service restarting, much like monit would on Linux.

As far as we can tell, some internal refactoring in how WinSW talks to
the Windows Service Manager resulted in this behavior change.  We
commented out the place where the Windows Service Manager is notified
that the wrapped service that exited 0 should be set to the stopped
state.

[#187232656]

Signed-off-by: Daniel Felipe Ochoa <daniel-felipe.ochoa@broadcom.com>