Attempts to make unmount more reliable
This commit executes the `fuser -k` with `nohup` in an attempt to ensure
that `fuser` survives the termination of `sshd`, which the comment
immediately after this line claims is a thing that happens.

We've somewhat-frequently seen in CI output very much like the following:

==========================================================
Remote Cmd Runner Running remote command '%s' sudo fuser -km /var/log
  Remote Cmd Runner Running remote command '%s' sudo umount /var/log
  CombinedOutput() FAILED TO EXECUTE: %s ERROR: %s sudo umount /var/log Process exited with status 32
  DetachDevice: Error running sudo umount /var/log: Process exited with status 32Remote Cmd Runner Running remote command '%s' sudo rm -rf /var/log
  CombinedOutput() FAILED TO EXECUTE: %s ERROR: %s sudo rm -rf /var/log Process exited with status 1
  [FAILED] in [SynchronizedBeforeSuite] - /tmp/build/5d05d134/bosh-agent/integration/integration_suite_test.go:36 @ 06/15/24 10:16:46.211
  << Timeline

  [FAILED] Unexpected error:
      <errors.ComplexError>:
      Error running sudo rm -rf /var/log: Process exited with status 1
      {
          Err: <*errors.errorString | 0xc0000364f0>{
              s: "Error running sudo rm -rf /var/log",
          },
          Cause: <*ssh.ExitError | 0xc00015cf40>{
              Waitmsg: {status: 1, signal: "", msg: "", lang: ""},
          },
      }
  occurred
  In [SynchronizedBeforeSuite] at: /tmp/build/5d05d134/bosh-agent/integration/integration_suite_test.go:36 @ 06/15/24 10:16:46.211
==========================================================

Our theory is that sometimes `fuser` kills the `sshd` that spawned the
shell that spawned `fuser`, which interrupts `fuser` in the middle of
killing processes. This would mean that some processes that had
/var/log open do not get killed, which results in the failure of the
umount of /var/log.

We'll see if this actually fixes anything. We expect that it can't
actually make things worse.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>