Lazily umount /var/log
This is another attempt at eliminating the failures to run DetachDevice
on '/var/log'.

Because
* umount is failing with a exit code of 32 (which either means
  "nothing is mounted there" or "mount point is busy"),
* Anything still using '/var/log' may get around to finishing up by the
  time we next remount '/var/log'
* It's not clear where '/var/log' is being mounted in the tests.
  It may be that it's being bind-mounted by the Agent, itself, which
  would mean that it's bind mounting something like
  '/var/vcap/data/root_log' to '/var/log'.
* We're later umounting '/var/vcap/data', so if we're bind-mounting
  something in there, and processes don't stop using it by that second
  unmount, then this commit should just push the failure a little further
  down the cleanup function, rather than quietly propagating an error
  state.

this seems to me to be a relatively safe thing to do.

However, IF we're bind-mounting '/v/v/d/root_log' to '/var/log', and we
later decide to ALSO lazily unmount `/var/vcap/data', then one potential
problem is that (at least for the 'ext' family of filesystem, and maybe
(all?) others) it looks like one can mount the same device in multiple
mount points without specifying the '--bind' flag. This means that any
processes that stay up and running will continue to write to this
filesystem, which could pollute future test runs.

The good news here is that if we're doing a 'mkfs' on the device that
houses the filesystem somewhere in test setup, the 'mkfs' WILL fail
pretty loudly if the device is mounted anywhere (even in a detached
mount), so we MAY be able to detect when processes that have that path
open just refuse to die.