Noble VMs on AWS don't have /dev/sdX devices for some reason.
We'd like to use the idDevicePathResolver and disk hints instead, but we don't
actually know the volume id for the ephemeral disk until the instance is created
at which point it's too late to put the disk into into the metadata because the
VM has to be in a stopped state to do that.

So instead, we just guess here and hope the nvme devices match up. Which isn't
much worse than what we were doing before, but AWS was also just sort of
hoping that the device would end up on /dev/sdb.