Ensure SystemTemp dir is set to /var/vcap/data/tmp
As of the March 2025 Windows Update, the C:\Windows\SystemTemp directory
is used by default as the temporary directory for system users. This
breaks BOSH lifecycle scripts that record large amounts of data in
temporary directories because it is on the root disk instead of the
expected ephemeral disk.

This commit explicitly sets the SystemTemp environment variable to
ensure that system users also use the temp directory on the ephemeral
disk.

For more information, see:
https://support.microsoft.com/en-us/topic/march-11-2025-kb5053596-os-build-17763-7009-d6bd4320-de45-4d63-ad47-b7c4789421c4#id0elbj=improvements
https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-gettemppath2a#remarks

We believe that job services created from the monit file are using temp
directories on the root disk. We have currently decided _not_ to change
this because we have not encountered problems with jobs themselves, only
lifecycle scripts. We suspect most jobs already explicitly point
themselves at `/var/vcap/data`, or otherwise use temp directories for
small files.

Signed-off-by: Chris Selzo <chris.selzo@broadcom.com>