Ignore SR-IOV VF interfaces in MAC Address Detector
When running Linux on cloud VMs with accelerated networking (SR-IOV),
it’s critical to distinguish synthetic (paravirtualized) network
interfaces from Virtual Function (VF) interfaces (direct hardware
virtualization).

On Azure, synthetic and VF interfaces share the same MAC address. It is
recommended to only use the synthetic interface to avoid disruptions
during events like Azure live-migrations.

To ensure the bosh-agent uses the correct NIC, the mac address detector
is changed to ignore NICs that have the alias "sriov-vf", which is
configured in the bosh stemcell when on Azure.

Resolves #366